# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/destinasi": {
      "get": {
        "operationId": "DestinasiController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar destinasi berhasil diambil"
          }
        },
        "summary": "Menampilkan daftar destinasi wisata",
        "tags": [
          "Destinasi"
        ]
      },
      "post": {
        "operationId": "DestinasiController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Destinasi berhasil dibuat"
          },
          "400": {
            "description": "Data tidak valid"
          }
        },
        "summary": "Menambahkan destinasi baru (khusus admin)",
        "tags": [
          "Destinasi"
        ]
      }
    },
    "/destinasi/{id}": {
      "get": {
        "operationId": "DestinasiController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Menampilkan detail destinasi",
        "tags": [
          "Destinasi"
        ]
      },
      "patch": {
        "operationId": "DestinasiController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Mengubah sebagian data destinasi (khusus admin)",
        "tags": [
          "Destinasi"
        ]
      },
      "delete": {
        "operationId": "DestinasiController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Menghapus destinasi (khusus admin)",
        "tags": [
          "Destinasi"
        ]
      }
    }
  },
  "info": {
    "title": "Wisataku API",
    "description": "API untuk platform informasi dan reservasi destinasi wisata",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika"
          },
          "kategori": {
            "type": "string",
            "example": "Pantai"
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000
          }
        },
        "required": [
          "nama",
          "kategori",
          "hargaTiket"
        ]
      },
      "UpdateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika"
          },
          "kategori": {
            "type": "string",
            "example": "Pantai"
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000
          }
        }
      }
    }
  }
}